LPS-X Provides a global env.d.ts file for ambient types
