+++ modules/apps/frontend-js/frontend-js-state-web/src/main/resources/META-INF/resources/deepFreeze.d.ts
@@ -11,17 +11,5 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-interface EventFacade {
-    currentTarget: any;
-    details: any;
-    relatedTarget: any;
-    target: any;
-    type: any;
-    halt(): void;
-    preventDefault(): void;
-    stopImmediatePropagation(): void;
-    stopPropagation(): void;
-    [key: string]: any;
-}
-export declare type EventCallback = (event: EventFacade) => void;
-export {};
+import type { Immutable } from './types';
+export default function deepFreeze<T>(value: T): Immutable<T>;
