+++ modules/apps/frontend-js/frontend-js-state-web/types/src/main/resources/META-INF/resources/State.d.ts
@@ -248,10 +248,3 @@ declare const State: {
  * ambient module declarations."
  */
 export default State;
-declare global {
-	interface Window {
-		Liferay: {
-			State: typeof State;
-		};
-	}
-}
