+++ modules/global-env.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/global.d.ts" />
+
+/// <reference path="apps/frontend-js/frontend-js-spa-web/src/main/resources/META-INF/resources/global.d.ts" />
+
+/// <reference path="apps/frontend-js/frontend-js-state-web/src/main/resources/META-INF/resources/global.d.ts" />
